requirements.txt
mcp==1.13.1
python-dotenv==1.1.1
rich==14.1.0